Skip LuaJIT detection if LuaRocks doesn't support it
This re-enables support for LuaRocks v3.1.x which was lost when LuaJIT
detection was added back in 961c204 (v0.2.0) because it assumed access
to an API that did not exist then. Everything else already still works
on that old version because only Lua 5.1 triggers this code path at all.
The ancient version of LuaRocks is significant because it features in
Ubuntu LTS releases.